requirements.txt
Flask
supabase-py
